Refactored some API into DataArrayUtilities so other filters can use the same codes
